+++ src/main/java/ru/vlsv/servlets/ProductsServlet.java
@@ -1,0 +1,53 @@
+package ru.vlsv;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import ru.vlsv.examples.BasicServlet;
+
+import javax.servlet.ServletException;
+import javax.servlet.annotation.WebServlet;
+import javax.servlet.http.HttpServlet;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.io.IOException;
+import java.io.PrintWriter;
+import java.sql.Connection;
+import java.sql.DriverManager;
+
+/**
+ * GeekBrains Java, JavaEEWebApp.
+ *
+ * @author Anatoly Lebedev
+ * @version 1.0.0 05.06.2020
+ * @link https://github.com/Centnerman
+ */
+
+@WebServlet(name = "ProductsServlet", urlPatterns = "/products")
+public class ProductsServlet extends HttpServlet {
+
+    private static final Logger logger = LoggerFactory.getLogger(ProductsServlet.class);
+
+    @Override
+    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+
+        resp.setContentType("text/html");
+        PrintWriter writer = resp.getWriter();
+        try {
+            logger.info("Connection succesful");
+            writer.println("Connection to ProductDB succesfull!");
+            resp.getWriter().write("Products catalog");
+            getServletContext().getRequestDispatcher("/products.jsp").include(req, resp);
+        } catch (Exception ex) {
+            writer.println("Connection failed...");
+            writer.println(ex);
+            logger.info("Connection failed. We`re get exception: " + ex);
+        } finally {
+            writer.close();
+        }
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        super.doPost(req, resp);
+    }
+}
